09MMS/requirements.txt
# 甯傚満寰粨鏋勪豢鐪熷紩鎿?(MMS) - Python渚濊禆鍖?
# NeuroTrade Nexus 寮€鍙戝洟闃?
# 鐗堟湰: 1.0.0
# 鍒涘缓鏃堕棿: 2024-12-01

# Web妗嗘灦
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 寮傛鏀寔
aiofiles==23.2.1
aiohttp==3.9.1
aiosqlite==0.19.0

# 鏁版嵁搴?
sqlalchemy==2.0.23
alembic==1.13.1

# 缂撳瓨
redis==5.0.1
aioredis==2.0.1

# 娑堟伅闃熷垪
pyzmq==25.1.2

# 鏁版嵁澶勭悊
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4

# 閲戣瀺鏁版嵁
yfinance==0.2.28
quandl==3.7.0
# alpaca-trade-api==3.1.1  # 鏆傛椂绉婚櫎锛屽瓨鍦ㄤ緷璧栧啿绐?

# 鏈哄櫒瀛︿範锛堝彲閫夛級
scikit-learn==1.3.2
statsmodels==0.14.1

# 閰嶇疆绠＄悊
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# 鏃ュ織鍜岀洃鎺?
loguru==0.7.2
prometheus-client==0.19.0

# 绯荤粺鐩戞帶鍜岄獙璇?
psutil==5.9.6
jsonschema==4.20.0

# 娴嬭瘯妗嗘灦
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-timeout==2.2.0

# 浠ｇ爜璐ㄩ噺
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.6.0

# 绫诲瀷娉ㄨВ
types-redis==4.6.0.11
types-PyYAML==6.0.12.12
types-requests==2.31.0.10

# HTTP瀹㈡埛绔?
requests==2.31.0
httpx==0.25.2

# 鏃堕棿澶勭悊
python-dateutil==2.8.2
pytz==2023.3

# 鏁板璁＄畻
sympy==1.12

# 搴忓垪鍖?
orjson==3.9.10
msgpack==1.0.3

# 鍔犲瘑鍜屽畨鍏?
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 鐜鍜岄儴缃?
gunicorn==21.2.0
docker==6.1.3

# 寮€鍙戝伐鍏?
ipython==8.18.1
jupyter==1.0.0
